fix: persist state between Cache/History and Network
